requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1
django-environ==0.11.2

# UI & Admin UI
django-jazzmin==3.0.1
django-admin-datta==1.0.17
django-crispy-forms==2.3
crispy-bootstrap5==2024.2

# Tools
djangorestframework==3.15.2
requests==2.32.3
pandas==2.2.3
graphviz==0.20.3
astor==0.8.1 

# Deployment
whitenoise==6.7.0
gunicorn==23.0.0

# DB
psycopg2-binary==2.9.9
psycopg2
#mysqlclient==2.1.1
django-dbbackup==4.2.1

# Widget tweaks
django-widget-tweaks==1.5.0
